requirements.txt
Flask
gunicorn
python-dotenv
openai>=1.30.0
numpy
pypdf
tqdm